mesh_database_client/requirements.txt
google_api_python_client==2.58.0
google_auth_oauthlib==0.5.2
googlemaps==4.6.0
numpy==1.23.1
pandas==1.4.3
protobuf==4.21.5
python-dotenv==0.20.0
